# Shortcut to avoid fmod