# Description
When benchmarking terms query (10000 terms) on long long type, I was supperised to find that the part costing most CPU is to judge whether terms have a decimal part because we execute a double mod for each long. So doing some check before mod may make sense.
![image](https://user-images.githubusercontent.com/52390227/103471437-6abe1f80-4dbb-11eb-8967-325d92b57385.png)
# Benchmark
I've done some jmh benchmark like this :
(`pct1[]` means 1% of 10000 objects will have a decimal part)
```
@Benchmark
@BenchmarkMode({Mode.Throughput})
@Fork(1)
@Measurement(iterations = 10)
@Warmup(iterations = 5)
public static void pct1(Blackhole blackhole) {
  for (int i=0;i<10000;i++) {
    blackhole.consume(hasDecimalPart(pct1[i]));
  }
}
```
**results**

  | baseline (ops/s) | candidate (ops/s)
-- | -- | --
1% decimal | 1224.221 ± 16.964 | 63735.415 ± 1682.500
50% decimal | 1166.410 ±  7.586 | 2235.981 ±   15.083
99% decimal | 1172.629 ± 58.190 | 1201.765 ±   32.000