@gf2121 2021-01-03
I've tried some other way to check a double's decimal part, and find that `Math.floor(doubleValue) != doubleValue` could be 50 times faster than a double mod. So i changed it in a new commit. 
Here is the benchmark result (check 10000 doubles per measurement):

```
Benchmark               Mode  Cnt      Score      Error  Units
MyBenchmark.baseline   thrpt   10   1217.505 ±   10.299  ops/s
MyBenchmark.candidate  thrpt   10  49250.338 ± 1069.798  ops/s
```